requirements.txt
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
asyncio==3.4.3
cachetools==5.3.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
DataRecorder==3.5.3
distro==1.9.0
DownloadKit==2.0.4
DrissionPage==4.1.0.4
easy-media-utils==0.1.5
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.16.1
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet
gspread==6.1.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.10
lxml==5.3.0
numpy==2.1.1
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.23.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
pillow
playwright==1.40.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core
pyee==11.0.1
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
StrEnum==0.4.15
tldextract==5.1.2
tqdm==4.66.2
typing_extensions==4.11.0
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
websocket-client==1.8.0
wheel==0.41.2
newspaper3k
feedparser~=6.0.11
textstat~=0.7.4
tranco~=0.8.1
lxml_html_clean
